+++ code/src/sixsq/nuvla/server/resources/deployment.clj
@@ -21,7 +21,8 @@ a container orchestration engine.
     [sixsq.nuvla.server.resources.resource-metadata :as md]
     [sixsq.nuvla.server.resources.spec.deployment :as deployment-spec]
     [sixsq.nuvla.server.util.metadata :as gen-md]
-    [sixsq.nuvla.server.util.response :as r]))
+    [sixsq.nuvla.server.util.response :as r]
+    [sixsq.nuvla.server.util.time :as time]))
 
 
 (def ^:const resource-type (u/ns->type *ns*))
@@ -33,8 +34,9 @@ a container orchestration engine.
 (def ^:const create-type (u/ns->create-type *ns*))
 
 
-(def collection-acl {:query ["group/nuvla-user"]
-                     :add   ["group/nuvla-user"]})
+(def collection-acl {:query       ["group/nuvla-user"]
+                     :add         ["group/nuvla-user"]
+                     :bulk-action ["group/nuvla-user"]})
 
 
 (def actions [{:name           "start"
@@ -233,8 +235,10 @@ a container orchestration engine.
         create-log-op       (u/action-map id :create-log)
         clone-op            (u/action-map id :clone)
         check-dct-op        (u/action-map id :check-dct)
+        fetch-module-op     (u/action-map id :fetch-module)
         upcoming-invoice-op (u/action-map id :upcoming-invoice)
         can-manage?         (a/can-manage? resource request)
+        can-edit-data?      (a/can-edit-data? resource request)
         can-clone?          (a/can-view-data? resource request)]
     (cond-> (crud/set-standard-operations resource request)
 
@@ -255,6 +259,8 @@ a container orchestration engine.
 
             can-manage? (update :operations conj check-dct-op)
 
+            (and can-manage? can-edit-data?) (update :operations conj fetch-module-op)
+
             (not (utils/can-delete? resource))
             (update :operations utils/remove-delete))))
 
@@ -267,6 +273,16 @@ a container orchestration engine.
              :nuvla/authn auth/internal-identity)
       (crud/edit)
       :body))
+
+;; FIXME: remove after this date:
+(def gnss-expiry-date "2021-08-01T00:00:00.000Z")
+(def gnss-groups ["group/gnss-admin" "group/gnss-user"])
+(defn gnss-group?
+  [request expiry-date]
+  (let [claim (auth/current-active-claim request)]
+    (and
+      (boolean (time/before? (time/now) (time/date-from-str expiry-date)))
+      (some #(= claim %) gnss-groups))))
 
 
 (defmethod crud/do-action [resource-type "start"]
@@ -288,8 +304,13 @@ a container orchestration engine.
           state          (if (= execution-mode "pull") "PENDING" "STARTING")
           new-deployment (-> deployment
                              (assoc :state state)
-                             (assoc :api-credentials (utils/generate-api-key-secret id
-                                                       (when data?
+
+                             (assoc :api-credentials (utils/generate-api-key-secret
+                                                       id
+                                                       (when (or data?
+                                                                 (gnss-group?
+                                                                   request
+                                                                   gnss-expiry-date))
                                                          (auth/current-authentication request))))
                              (cond-> subs-id (assoc :subscription-id subs-id))
                              (edit-deployment request))]
@@ -433,6 +454,34 @@ a container orchestration engine.
   [resource]
   (utils/get-context resource false))
 
+
+(def bulk-action-impl (std-crud/bulk-action-fn resource-type collection-acl collection-type))
+
+(defmethod crud/bulk-action [resource-type "bulk-update"]
+  [request]
+  (bulk-action-impl request))
+
+
+(defmethod crud/do-action [resource-type "fetch-module"]
+  [{{uuid :uuid} :params body :body :as request}]
+  (let [id          (str resource-type "/" uuid)
+        deployment  (crud/retrieve-by-id-as-admin id)
+        module-href (:module-href body)]
+    (a/throw-cannot-edit deployment request)
+    (when (or (not (string? module-href))
+              (str/blank? module-href)
+              (not (str/starts-with? module-href (-> deployment
+                                                     (get-in [:module :href])
+                                                     (str/split #"_")
+                                                     first))))
+      (throw (r/ex-response "invalid module-href" 400)))
+    (let [authn-info  (auth/current-authentication request)
+          module      (utils/resolve-module (assoc request :body {:module {:href module-href}}))
+          dep-updated (update deployment :module utils/merge-module module)]
+      (crud/edit {:params      {:uuid          uuid
+                                :resource-name resource-type}
+                  :body        dep-updated
+                  :nuvla/authn authn-info}))))
 
 ;;
 ;; initialization