+++ code/test/sixsq/nuvla/server/resources/callback_deployment_update_test.clj
@@ -77,9 +77,9 @@
                               session
                               (content-type "application/json"))
          session-admin    (header session-anon authn-info-header
-                                  "group/nuvla-admin group/nuvla-user group/nuvla-anon")
+                                  "group/nuvla-admin group/nuvla-admin group/nuvla-user group/nuvla-anon")
          session-user     (header session-anon authn-info-header
-                                  "user/jane group/nuvla-user group/nuvla-anon")
+                                  "user/jane user/jane group/nuvla-user group/nuvla-anon")
 
          ;; setup a module that can be referenced from the deployment
          module-id        (-> session-user