+++ code/test/sixsq/nuvla/server/resources/callback_lifecycle_test.clj
@@ -28,7 +28,7 @@
                           (content-type "application/json"))
         session-admin (header session authn-info-header
                               "group/nuvla-admin group/nuvla-user group/nuvla-anon")
-        session-user  (header session authn-info-header "user/jane group/nuvla-user group/nuvla-anon")
+        session-user  (header session authn-info-header "user/jane user/jane group/nuvla-user group/nuvla-anon")
         session-anon  (header session authn-info-header "group/nuvla-anon")]