+++ code/test/sixsq/nuvla/server/resources/cloud_entry_point_lifecycle_test.clj
@@ -27,8 +27,8 @@
   (let [session-anon  (-> (ltu/ring-app)
                           session
                           (content-type "application/json"))
-        session-admin (header session-anon authn-info-header "group/nuvla-admin group/nuvla-user group/nuvla-anon")
-        session-user  (header session-anon authn-info-header "user/jane-updater")]
+        session-admin (header session-anon authn-info-header "group/nuvla-admin group/nuvla-admin group/nuvla-user group/nuvla-anon")
+        session-user  (header session-anon authn-info-header "user/jane user/jane-updater")]
 
     ; retrieve root resource (anonymously should work) and verify schema
     (let [cep (-> session-anon