+++ code/test/sixsq/nuvla/server/resources/credential_infrastructure_service_amazonec2_lifecycle_test.clj
@@ -26,9 +26,9 @@
   (let [session               (-> (ltu/ring-app)
                                   session
                                   (content-type "application/json"))
-        session-admin         (header session authn-info-header "group/nuvla-admin group/nuvla-user group/nuvla-anon")
-        session-user          (header session authn-info-header "user/jane group/nuvla-user group/nuvla-anon")
-        session-anon          (header session authn-info-header "user/unknown group/nuvla-anon")
+        session-admin         (header session authn-info-header "group/nuvla-admin group/nuvla-admin group/nuvla-user group/nuvla-anon")
+        session-user          (header session authn-info-header "user/jane user/jane group/nuvla-user group/nuvla-anon")
+        session-anon          (header session authn-info-header "user/unknown user/unknown group/nuvla-anon")
 
         name-attr             "name"
         description-attr      "description"