+++ code/test/sixsq/nuvla/server/resources/data_record_key_lifecycle_test.clj
@@ -46,7 +46,7 @@
                           (content-type "application/json"))
         session-admin (header session-anon authn-info-header
                               "group/nuvla-admin group/nuvla-user group/nuvla-anon")
-        session-user  (header session-anon authn-info-header "user/jane group/nuvla-user group/nuvla-anon")]
+        session-user  (header session-anon authn-info-header "user/jane user/jane group/nuvla-user group/nuvla-anon")]
 
     ;; create namespace
     (-> session-admin