+++ code/test/sixsq/nuvla/server/resources/group_template_lifecycle_test.clj
@@ -24,9 +24,9 @@
   (let [session       (-> (ltu/ring-app)
                           session
                           (content-type "application/json"))
-        session-admin (header session authn-info-header "group/nuvla-admin group/nuvla-user group/nuvla-anon")
-        session-jane  (header session authn-info-header "user/jane group/nuvla-user group/nuvla-anon")
-        session-anon  (header session authn-info-header "user/unknown group/nuvla-anon")]
+        session-admin (header session authn-info-header "group/nuvla-admin group/nuvla-admin group/nuvla-user group/nuvla-anon")
+        session-jane  (header session authn-info-header "user/jane user/jane group/nuvla-user group/nuvla-anon")
+        session-anon  (header session authn-info-header "user/unknown user/unknown group/nuvla-anon")]
 
     ;; admin user collection query should succeed and contain exactly 1 template
     (-> session-admin