+++ code/test/sixsq/nuvla/server/resources/hook_reset_password_lifecycle_test.clj
@@ -27,8 +27,8 @@
   (let [reset-link    (atom nil)
         app           (ltu/ring-app)
         session-json  (content-type (session app) "application/json")
-        session-anon  (header session-json authn-info-header "user/unknown group/nuvla-anon")
-        session-admin (header session-json authn-info-header "group/nuvla-admin group/nuvla-user group/nuvla-anon")]
+        session-anon  (header session-json authn-info-header "user/unknown user/unknown group/nuvla-anon")
+        session-admin (header session-json authn-info-header "group/nuvla-admin group/nuvla-admin group/nuvla-user group/nuvla-anon")]
 
     (with-redefs [email-utils/extract-smtp-cfg
                                       (fn [_] {:host "smtp@example.com"