+++ code/test/sixsq/nuvla/server/resources/session/utils_test.clj
@@ -74,8 +74,8 @@
 
   (let [app              (ltu/ring-app)
         session-json     (content-type (session app) "application/json")
-        session-admin    (header session-json authn-info-header "group/nuvla-admin group/nuvla-user group/nuvla-anon")
-        session-anon     (header session-json authn-info-header "user/unknown group/nuvla-anon")
+        session-admin    (header session-json authn-info-header "group/nuvla-admin group/nuvla-admin group/nuvla-user group/nuvla-anon")
+        session-anon     (header session-json authn-info-header "user/unknown user/unknown group/nuvla-anon")
         session-user     (header session-json authn-info-header "user group/nuvla-user")
 
         href             (str st/resource-type "/password")