+++ code/test/sixsq/nuvla/server/resources/user_minimum_lifecycle_test.clj
@@ -32,9 +32,9 @@
         session           (-> (ltu/ring-app)
                               session
                               (content-type "application/json"))
-        session-admin     (header session authn-info-header "group/nuvla-admin group/nuvla-user group/nuvla-anon")
-        session-user      (header session authn-info-header "user/jane group/nuvla-user group/nuvla-anon")
-        session-anon      (header session authn-info-header "user/unknown group/nuvla-anon")
+        session-admin     (header session authn-info-header "group/nuvla-admin group/nuvla-admin group/nuvla-user group/nuvla-anon")
+        session-user      (header session authn-info-header "user/jane user/jane group/nuvla-user group/nuvla-anon")
+        session-anon      (header session authn-info-header "user/unknown user/unknown group/nuvla-anon")
 
         template          (-> session-admin
                               (request template-url)
@@ -152,7 +152,7 @@
 
           email                (get-in href-create [:template :email])
 
-          session-created-user (header session authn-info-header (str user-id " group/nuvla-user group/nuvla-anon"))
+          session-created-user (header session authn-info-header (str user-id " " user-id " group/nuvla-user group/nuvla-anon"))
 
           {user-acl  :acl
            email-id  :email